requirements.txt
streamlit
pillow
exifread
numpy
opencv-python-headless